[FIX] knowledge: fix embedded views renaming
Purpose:
--------
Commit [1] refactored the breadcrumbs. Since then, renaming an embedded
view did not show the new name until the view was reloaded (because
the active element has not the same class as before).

This commit uses the new class of the active element to update the name
of the embedded view.

[1]: https://github.com/odoo/odoo/commit/caef16ee4e5351142ef04f02798999ccb38a56bc

Task-3329989

closes odoo/enterprise#42084

X-original-commit: 6c35133b2a210fd75a79ec68b88cb2746b00fc5d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>