[FIX] account_reports: ensure correct rates in multicurrency revaluation
Currently, in multicurrency revaluation, the currency rates are
normalized on to the company's currency.

However, the normalization can be incorrect if the company's currency
appeared first in the rates dictionary. This is due to the fact that the
company's exchange rate value can be modified before other currencies
are normalized.

Part-of: odoo/enterprise#41173