[FIX] project_timesheet_forecast: generate timesheets for flexible res.
Current behaviour:
When generating the timesheets from the planning app with a resource
that is set with flexible hours, the generate time duration for the
timesheet duration aren't correct.

Expected behaviour:
The time of the generated timesheets should follow the same logic as
when we compute the `allocation_hours` of the slot.

Steps to reproduce:
- Install Sales, project_timesheet_forecast
- Sell a service, make sure it generates a project and task.
- Check Mitchell's flexible hours
(Employees > Mitchel Admin > Work Information > Flexible Hours)
- In Planning, create a new slot, assign the SO and the service we sold
- Set the date of the slot from 6AM to 18PM (in a day at least 1
  month in the past), and overwrite the allocation hours to 12h.
  (should have 150% allocation percentage being recomputed).
- Publish the shift
- In the list view, select the slot, and from the action menu >
  "Generate Timesheets"
- Timesheets, observe the generated timesheet is 11.4, instead of
  12h.

Reason for the problem:
In the `_action_generate_timesheet` flexible resources weren't
handles at all. So we were using the resource's calendar and
computing based on work interval, as we would with an employee that is
not flexible. So for a 38h/w calendar (taken from the contract that
was active during the period we've put the slot), the computed
timesheet is `7.6h * 150% = 11.4h`, instead of `12h`.
So the issue appears only for people that recently changed contracts
work hours.

Fix:
Branch the routine so it handles flexible resources as by specs,
so it's inlined with the `allocation_hours` of the slot.

Affected versions:
- 16.0
- saas-16.1
- saas-16.2
- saas-16.3
- master

opw-3281910

closes odoo/enterprise#41194

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>