[FIX] l10n_ec_edi: Document type supported for tax support in tests
(i)   The tax support named '01' doesn't support document types with code '09'. Use document types with code '01' by default in withhold tests to better reflect real scenarios.
(ii)  Changed partner b to a because its identification type is RUC and covers most of the real scenarios that withholds can be generated, so it is suggested to keep most of the unit tests with this partner.
(iii) For withholds from foreign partners specifically, the tax support should be '02' and not '01'. The related test is fixed and improved:
- Partner b is kept instead of the new default partner, because this partner is used for unit tests with real scenarios of local people with VAT or Passport identification type and foreign people.
- Partner b is adapted to use 'vat' identification because it is a document for a foreign partner and the XML must show certain tags such as tipoSujetoRetenido, which only appears when the identification type code in tipoIdentificacionSujetoRetenido tag is '08' and that was not currently being displayed by the unit test.
- Custom withhold lines are created with appropriate taxes/values

X-original-commit: 6a9fa8a1d1f3737ddd5c04b33324f47d513c4668
Part-of: odoo/enterprise#41990