[FIX] documents: fix issue when drag & drop document within workspace
Before this commit:
When we drag and drop the existing file in the workspace. The file is duplicated.

Reason:
Before version saas-16.3 'ev.dataTransfer.types' returns types 'Files' only
when we drag & drop a new file in the workspace.
But from version saas-16.3 the types 'Files' return for both cases.

Case 1:- drag & drop a new file
Case 2:- drag and drop an existing file.

So the existing condition fails which causes an issue.

After this commit:
Now we check for the key, which is only present in existing records.
As a result, the existing documents are no longer duplicated.

Task-3299263

closes odoo/enterprise#44889

X-original-commit: 560cba279f8bb7f2cfe16d86eca2c9f6f2c22077
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>