[ADD] l10n_fr_reports: add French EDI VAT
Allow to send the VAT report to the government, using the API of ASPOne.
Add a new model `account_report_async_export` to store the exports of a
report and process them using a CRON.

The flow for the french EDI VAT is as follows:

1. Open a wizard from the French VAT report

2. When confirming it, an account.report.async.export containing an xml
is created. The xml is also sent to ASPOne.

3. The list of account.report.async.export is available from the
"Reporting" > "EDI exports" menu

4. The CRON will process the pending account.report.async.export

For ASPOne, it's done in 2 phases. First, by calling
`getInterchangesByDepositID`, then by calling `getDeclarationDetails`.
Note that if attachments were generated (i.e. edifact files), they can
be retrieved using `getRecipientReports_v2`.

task-2673918

closes odoo/enterprise#41264

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>