[FIX] web_studio: fix undeterministic tests (x2many subviews)
The refactoring of studio's ViewEditorManager (d6d8aeff58e46373f72fc932cd84822a29ea76fb) implied
some adaptations in tours.

One of the patterns used was, to test instances of navigation down a x2many path, to wait
in the DOM for the studio breadcrumb to update.
This was largely insufficient as the breadcrumb updates within the next frame, but the requested
view (the x2many [list|form] view) took a little longer to load and display.

This created underterministic errors at several places before this commit.

This commit changes this wobbly pattern into something more certain: we wait for the wanted
view to be in the DOM before doing editions on it.

runbot-error-20435
runbot-error-20239
runbot-error-20843

closes odoo/enterprise#41282

X-original-commit: 2032f460f55fdfa8b7210dd44a0b6d832654914a
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>