[FIX] timesheet_grid: specify the widget
Issue:
In debug mode, on the form view of an employee,
if we click on the "Timesheets" button,
an error appears.

Cause:
We do not call the right widget,
because none is specified in the view.
The default `Many2OneField` will be used,
but the props will not match.

Solution:
We have to call `TimesheetGridMany2OneField`
which is an extend of `Many2OneField`.

opw-3329404

closes odoo/enterprise#41670

X-original-commit: 04e97e7a35f51bd1bb355aceb9177f79ae1bf328
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>