[IMP] knowledge: adds comment system
This commit adds a brand new system to Knowledge, which is the ability to
add comments to a specific article.

In order to do that, we are using the `mail.thread` mixin with a brand new
model `knowledge.article.thread` which is used to keep every comment as its
own thread.
This model keeps track of its state, it can be resolved and unresolved.
A resolved thread is a thread that is not directly visible in the editor and
is considered as closed: no message should be sent on this thread.
If a thread has been resolved for longer than 1 month, it is removed from the
DB in order to not clog it with stale threads.

In order for user the user to create and handle threads we created 3 Component:
 * `KnowledgeCommentsThread`
 * `KnowledgeCommentsHandler`
 * `KnowledgeArticleCommentsPanel`

`KnowledgeCommentsThread` represents a singular comment thread inside the article
and is used by both the Panel and the Handler.
Each Comment uses a Thread and a Composer that interacts using a thread object.
You can send new comments, react to messages, reply to a specific one or edit
one of your own.
When using a smaller UI the box is replaced by the face of the last person who
commented in the thread. When clicking on it you open a popover that contains
the same elements as the box.

`KnowledgeCommentsHandler` is the link between the editor , the panel, the DB
and the comment system. It handles the creation and destruction of comments,
the changing of state for each comment. It only displays unresolved threads
but keeps track of all the comments.

`KnowledgeArticleCommentsPanel` is a panel similar to the ChatterPanel or the
PropertiesPanel that displays the comments following 3 modes: all, unresolved
and resolved.
You can interact inside the panel like you would with the handler, you can
answer comments, resolve them, etc. but you can also see unresolved comments
and reopen them to relaunch a discussion on a specific part of the article.

task-3317056

closes odoo/enterprise#41951

Related: odoo/odoo#127380
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>