[IMP] sale_subscription: pending payments warning in portal
Before this commit, when a customer had payed a subscription with a method that was not instantly confirmed (pending), it wasn't shown any information about the payment that was just made when going back to the subscription. Now, it is shown an alert informing about the date when the pending payment was started and another alert notifying that there is a pending payment.

closes odoo/enterprise#39542

Task-id: 2848272
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>