[FIX] web: Fix grid renderer row height
Before this commit, the grid row height has been set to be automatically
computed by the browser (the grid-template-rows property value was set
to "auto"). While this is a more maintainable approach, it messes up
with the virtual hook (`@web/core/virtual_hook`) which needs to know the
exact height of each row.

This commit ensures that the grid renderer uses its hard-coded row
height in its grid layout.

Part of task 3336233

closes odoo/enterprise#41595

X-original-commit: 38f39c76f0e85ad1157f4b912adcb83bf6ad3ad2
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>