[FIX] web_studio: calendar editor constrain modes according to scales
Have a calendar view with its attribute set to some value (eg `"month,year"`)
In studio, change the mode to day.

Before this commit, there was a crash because the calendar scales attribute constraints
the value of the mode and of the default mode.

After this commit, we don't even propose modes that are not part of the scales attribute.

opw-3316186

closes odoo/enterprise#41294

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>