[FIX] social, social_*: adapt post view layout
Impacted modules: social, social_crm, social_sale

Prior to this commit, the `post-stats` part of the post view was broken
since the font size was increased in Milk.
This commit fixes and improves the layout to work in all cases,
depending on the language used and the amounts displayed.

Also, this commit fixes a missing border when the post contains
attachments.

task-3335076
Part of task-3326263

closes odoo/enterprise#41778

X-original-commit: e7e854594f3ceadd94d0cc9d53958e03898e6fa9
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>