[IMP] web_studio: restore active tab and selected element
This commit improves the behavior of the Form editor when editing an
element contained inside a notebook page. Before this commit, whenever
an attribute or option is modified, the form renders with the first page
of the notebook visible, even when the modified element was located in
the second page.

Now, when editing an element in the second page, the form view renders and
automatically display the appropriate page with the element selected and
ready to be edited again.

A test has been added to verify that an edited element is still selected
after such attribute change.

task-3259495

closes odoo/enterprise#41692

Signed-off-by: Dardenne Florent (dafl) <dafl@odoo.com>