[FIX] account_asset: loss acount on asset sell wizard
Before this commit, the salvage value of the asset
was not taken into account when computing the
gain_or_loss field on the sell wizard. Therefore,
in some cases, the user was not able to select a
loss account, leading to miscalculation of the
lines values when creating the disposal move if no
loss account was set on the company (this is the
case when no disposal move has already been created
before).

Steps:

- Make sure there is no existing disposal move
  of any asset
- Create an asset for 1500$, with 1000$ not depreciable,
  in the past in order to have it liquidated
- Open the 'sell or dispose' wizard and dispose
  the asset
-> UserError: "Cannot create unbalanced journal entry"

opw-3325111

closes odoo/enterprise#41427

X-original-commit: 83bc6ba681a31bc75236dc5466d52a80c93fd020
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>