[IMP] pos_preparation_display: improve offline popup
Before when the point_of_sale was offline, popups were opened at each
RPC error. This was inconvenient for the user as the pos must be usable
offline.

Now, the error popup has been replaced by an alert popup explaining that
the point of sale works in limited version as long as it is offline.
The warning noise has also been removed.

closes odoo/enterprise#41323

Related: odoo/odoo#120826
Signed-off-by: Guilliams Adrien (adgu) <adgu@odoo.com>