[IMP] documents_spreadsheet: automatically removes empty spreadsheet
Task Description

When creating a new spreadsheet and directly closing it without any
changes, it is automatically saved, leading clients to have lots of
empty documents in the database and forcing them to regularly filter
theirs documents to archive empty/unused spreadsheets.
This task add an automated task in the autoremove cron to check every
day if there is any empty spreadsheet (without any changes) created
during the day before and, if there is some, to automatically remove it.

Related Task

task-2871820

closes odoo/enterprise#41333

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>