[FIX] account_accountant: Fix phantom exchange diff line in bank rec widget
- Create a bank transaction of 3000 in company currency (make sure the rate is 3:1 at this date).
- Create an invoice line of 2000 using a foreign currency with a 2:1 rate.
- Mount the invoice line into the widget => an exchange diff of 1000 has been created.
- Remove the invoice line
=> An open balance line is computed with 2000 instead of 3000.

When removing a 'new_aml' line in a multi-currency environment, the exchange difference
line wasn't removed so the open balance is computed by considering the exchange difference line (here 2000).
When computing lines_widget, the exchange diff are re-order to be added right after the targeted aml.
Since this line no longer exists, the exchange diff is not added and then, wasn't part of the final results.
So, the results was only an open balance with a wrong amount.

closes odoo/enterprise#41346

Task: 3336436
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>