[FIX] mail: show more options in chat window header
Before this commit, chat window header showed all actions, which made it quite bloated. The name of conversation was sometimes completely truncated by available actions.

This commit fixes the issue by showing at most 2 actions immediately in the chat window header. When there are more actions, it shows an expandable more menu.

As a consequence of this change, when a panel is open, it was harder to see the button to close the active panel (toggling off the button).
To solve this issue, when a panel is open in chat window, there's a floating button in the top-right corner of content to close the panel.

Task-3334793

closes odoo/enterprise#41466

X-original-commit: a6e6d928a4297eebee38208804875e391e81c095
Related: odoo/odoo#122419
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>