[FIX] website_sale_renting: fix date consistency issues
Currently, when a customer clicks on a renting product name, while in
the cart, he is redirected to the product page. The date picker should
show the dates of the sale order's first sale order line but it fails
and shows the wrong date because they are cached on the page.

Now, the date picker shows the date of the sale order's first sale order
line.

task-3293848

Part-of: odoo/enterprise#40534