[FIX] web_grid: prevent cell toggle when clicking the magnifier glass
Current behaviour:
In the timesheet's grid view, if we set the timesheet's UOM to
`Days/Half days` in the settings, the cells are toggle buttons that
cycle between 0 -> 0.5 -> 1 -> 0 ... When we click on the magnifier
glass on a cell, which should open a new view and show the analytic
lines related to that day, a click on the cell goes through, editing
the content of the cell.

Expected behaviour:
When we click on the magnifier glass, the cell shouldn't be edited.

Steps to reproduce:
- Install Timesheets
- Timesheets App > Configuration > Settings > change the encoding
  method to `Days/Half days`
- Go the Timesheet's app main grid view
- Click on a magnifier glass of one of the cells (for example a
  cell that has 0 as value).
- A new list view opens, but the value is 0.5, when you go back via
  the breadcrumbs, you can see the cell's value has changed to 0.5,
  should have stayed 0.

Reason for the problem:
There are 2 `onClick` listeners, 1 on the cell component, another one
the magnifier glass component, which is inside the cell component.
So when we press the magnifier glass, it's `onClick` is triggered,
then the `click` event *bubbles* up the components, and triggers the
`onClick` of the cell component, which updates the content of the cell.

Fix:
Use the target of the click event to know on what we clicked and trigger
the `onChange` based on that.

Affected versions:
- saas-16.2
- saas-16.3
- master

opw-3325115

closes odoo/enterprise#42266

X-original-commit: c3fe23dc71950c00c6b387a3f309c2e843951513
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>
Co-authored-by: xavierbol <xbo@odoo.com>