[FIX] account_reports: stop error when using horizontal groups w/o COA
Setup:
Have a DB with `account_accountant` installed and a company that does
not have a chart of accounts.

Steps to reproduce:
Set up a horizontal group on the General Ledger report with the domain
"Account"

If you try to select that horizontal group in the General Ledger, you
get a traceback.

opw-3164062

closes odoo/enterprise#41399

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>