[FIX] account_reports: stop error when using horizontal groups w/o COA
Setup:
Have a DB with `account_accountant` installed and a company that does
not have a chart of accounts.

Steps to reproduce:
Set up a horizontal group on the General Ledger report with the domain
"Account"

If you try to select that horizontal group in the General Ledger, you
get a traceback.

opw-3164062

closes odoo/enterprise#41463

X-original-commit: c640208d62d92de38ce7950c0f0305bb703f4c7f
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Nshimiyimana Serge Séna (sesn) <sesn@odoo.com>