[FIX] web_studio: changing modifiers: overwrite attrs if present
This commit fixes an issue that happened when one wanted to override
the attrs of a field (or button) with a static modifier while it had a
domain in its attrs originally. This makes sure that the attrs are
properly replaced by an empty object if needed to cancel the domain
and use the static modifier instead.

Part-of: odoo/enterprise#41066