[FIX] knowledge: display aside hidden buttons when no cursor
Currently, the mobile users can not see or reveal the buttons of the
sidebar for creating new articles or for joining a hidden article. The
user needs to hover those elements with a mouse pointer to reveal them.
In order to ensure that mobile users can interact with those buttons,
this commit will ensure that they will remain visible when the user does
not have a mouse pointer.

The button for joining a hidden article has different styling compared
to the button right next to it. This commit will update the styling of
that button to ensure visual consistency.

task-3297985

closes odoo/enterprise#41749

X-original-commit: 2a1c440f87850f54c939d1ff599c977e5cfbfc00
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>