[IMP] planning: display split tool on planning pills
planning.slot - Gantt view: When using using the Gantt view in Planning,
it is in some case necessary to split an existing shift that span across
hours/days/weeks/months. There is currently no simple way to do that.

This commit introduces a split tool in the Gantt view of planning,
allowing to simply cut planning slots. The tool is displayed when the
user hover a given pill at the junction between hours, days, weeks or
months depending on the scale of the Gantt view. Once clicked, the slot
is automatically split into two slots finishing and starting at the
junction selected.

If the records are grouped by resources in the Gantt view, a check on
the availability of the resource assigned to the shift will be performed
among splitting. If possible, the two resulting shifts will be shrunk
from the cutting line, to match the resource availability.

The split tool is not available in mobile view.

Other changes introduced by this commit are:
- cleaning of the file planning_security.xml by removing a useless
record rule
- recurrence is not copied anymire when duplicating a shift (reversing a
change introduced in https://github.com/odoo/enterprise/pull/23571)

task-3330231

closes odoo/enterprise#41516

Related: odoo/upgrade#4857
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>