[FIX] helpdesk: display Accept Emails From before alias_contact field
Issue:
The label 'accept emails from' is missing from the alias_contact field.

Cause:
The string applied to the field is not visible because it is inherited through
another model. Therefore, the string is not functioning as intended.

Solution:
To resolve this issue, a label needs to be applied to the field.

task-3318851

closes odoo/enterprise#41500

Related: odoo/odoo#122485
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>