[FIX] hr_payroll_account: prevent traceback when violates not-null constraint
When a user deletes the hr_salary_rule `House Rent Allowance` which is present
in the `hr_payroll_account` module. while the user reloads the demo data of the
`hr_payroll_acocunt` module at that time traceback will be generated in the log.

Steps to reproduce :
1. Install the `hr_payroll` module.
2. Open the payroll app.
3. Cancel all those pay_slips whose `House Rent Allowance` is in the
    pay_slip_line.
4. Delete all those pay_slips whose `House Rent Allowance` is in the
    pay_slip_line.
5. Select Configuration > Rules.
6. Delete the `House Rent Allowance` rule in the `hr_salary_rule` model.
7. Install the `hr_payroll_acocunt` module.
8. Install the `l10n_be` module.
9. Traceback will be generated in the log.

Applying this commit will solve this issue.

sentry - 4186952560

closes odoo/enterprise#42153

X-original-commit: 2751e10c309a5402f2474e3e6887a2a7dee5b6a2
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>