[FIX] web_enterprise: get correct mimetype for menu item icon
Since changes made in https://github.com/odoo/enterprise/pull/41117 we
cannot provide icons for menus in an other format than png. As it was
allowed before to use SVG, we don't want to restrict to only one format.

Before, it was trying to gess the mimetype based on the image content
which is not also the best case.

So as the icon is a binary field store=True, we can just take the
mimetype from the attachment.

closes odoo/enterprise#41529

X-original-commit: 88e29271d83fe85d10a01641c50909f55a9c4e9d
Related: odoo/odoo#122580
Signed-off-by: Masereel Pierre <pim@odoo.com>