[IMP] account_accountant: prevent foreign currency userError with domain
When creating a bank transaction (account.bank.statement.line), the user is able to choose the journal currency which then results in a user error. By adding a domain to the form and list views, we prevent the user from choosing the journal or company currency, thereby eliminating the user error.

Task-3338517

closes odoo/enterprise#41532

X-original-commit: 9b3dc1e1b39e0022aed2269038317e971f6721dc
Related: odoo/odoo#122595
Signed-off-by: Laurent Smet <las@odoo.com>