[FIX] sale_planning: corractly copy previous week when allocated percentage
Issues:
\* `_init_remaining_hours_to_plan` wasn't taking into account the
correct amount of hours planned, because it is a compute stored, and so
the compute method wasn't triggered.

\* in `_update_remaining_hours_to_plan_and_values`,
`remaining_hours_to_plan` was updated, taking the percentage into
account only in some cases. Yet we always want that.

\* in the same method, `remaining_hours` wasn't maxed, so it was
possible to create a shift with a much greater duration than the one it
was copied from.

\* in the same method, we the fact that we allowed `allocated_hours`
to stay in the `values` caused the other values to be recomputed when
the shift was created, whereas it should be the opposite.

Also:
\* `float_compare` can only return 3 values, so checking its return
value can be done with `==` and `!=` i/o `<` or `>`.

task-3251696

closes odoo/enterprise#43268

X-original-commit: 21fe7a0e39ef9047a73d7c3a258387d6b15073fb
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>