[FIX] account_taxcloud: prevent traceback when Taxcloud Category code is string
When user input some string value in 'Taxcloud Category' in product form view.
As the field 'Taxcloud Category' of 'code' field is integer as it is given as a
'rec_name'.

It will give an error with the message -

'invalid literal for int() with base 10'

Steps to Produce:-

1. Install 'account_taxcloud' module
2. Go to 'invoicing' module
3. Go to 'Vendors' -> 'Products'
4. Click on any product or create new
5. In 'TaxCloud Category' field input any string
6. Press enter

Trace-back will be generated.

Applying these changes will resolve this issue.

sentry - 3960203288

closes odoo/enterprise#41551

Signed-off-by: William André (wan) <wan@odoo.com>