[FIX] l10n_be_hr_payroll: fix action_validate rights in leave
Before this commit, an employee with no Payroll rights that wanted
to confirm a time off for an employee for which they're set as
the time off approver, and if the time off type approbation is set
to "Employee's Approver" only, the user would get an access error
due to rights on work entries.
This commit fixes that issue.

task-3329557

closes odoo/enterprise#42208

Signed-off-by: Kevin Baptiste <kba@odoo.com>