[FIX] project_sale_subscription: renewal did not created tasks or projects
Before this commit, when a product was configured to create tasks or projects
it worked only for the first subscription. When the order was upsell or renewed,
it would not generate any new task or project.

This commit makes sure that upsell do not generate service objects because the
pro-rata could not be handled (50% of a task created) but renewal should definitely
create new objects.

Taskid: 3339544
X-original-commit: 96582f95f2b918dd64814df9d37f82a3a4888ee5
Part-of: odoo/enterprise#44550