[FIX] mrp_subcontracting_account_enterprise: fix cost report
This fix aims to allows you to open the cost analisys on BOMS which are configured as
subcontracting.
Before the fix, if you had a subcontracting purchase order for which you sent all the components
but only received a partial amount fo the finished product, a back order would be created. This
would in turn create a second MO (which is hidden as its subcontracted). Once all the products
received, you end up with two stock moves and two manufacturing orders.
The fetch of the unit cost for the report would send back an 'expected singleton' traceback as
there are multiple stock moves for the initial production.
With this fix we loop through the production_ids that are given in paramater of the function and
the stock moves as well, out of safety (even though there is already a filter to avoid byproducts -
 you never know).
OPW-3318933

closes odoo/enterprise#41886

X-original-commit: a75dd6201cc87e9148323b4b2d425a5c778b4aa8
Signed-off-by: William Henrotin (whe) <whe@odoo.com>