[IMP] industry_fsm: display project color in calendar
In this commit, we have implemented the functionality to show the color
of the project in the left side panel of the calendar and calendar entries.
This feature is particularly useful when the 'worksheets' functionality
is disabled, as we cannot rely on worksheet templates in such cases.

task-3251766

closes odoo/enterprise#41569

Related: odoo/odoo#130700
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>