[FIX] website_sale_renting: correct display zero price rental product
When the "Prevent Sale of Zero Priced Product" parameter is activated,
a rental product with a zero price period will always its information
displayed.

With this commit, the display will be the same as for a product with
a sale price of zero, i.e. just the message to warn that it is
not available and a "Contact Us" button.

opw-3323271

closes odoo/enterprise#41535

Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>