[FIX] stock: don't set draft move with 0 quantity
Usecase:
- Create a picking and a move without quantity
- Click or the show detail or save

The picking goes into the state draft while it shouldn't.
It happens because when a quantity done is set, it will
goes to process_increase that will assign the move. With
0 qty, it's not assign and stay in draft (default value)
The picking state is then computed based on the move state.

To fix it, force the state to assign on stock.move in immediate
transfer without quantity.

closes odoo/enterprise#41548

Task: 3256447
Related: odoo/odoo#122445
Signed-off-by: William Henrotin (whe) <whe@odoo.com>