[IMP] l10n_de_pos_cert: replace rendering of json object from the xml
Currently, the json object needed to close the session is obtained from an xml template that is transformed in python.

This approach is not necessary and is also error prone. In this PR I refactor the code in order to create the object directly in Python.

Task 3323623

Part-of: odoo/enterprise#41568