[FIX] spreadsheet_edition: odoo menu link issue
When you set a menu in the prompt dropdown and then try to set another menu
in the same prompt, you can open the dropdown but the new selection is not taken
into account even if you clear the first menu.

This commit solves the issue by taking into account, the id of selectedMenu
instead of props.menuId.

Also, if condition for `updateWidget` is altered to ensure that dropdown
still opens when input is cleared.

Task ID : 2821480

closes odoo/enterprise#41948

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>