[FIX] account_reports: two closing entries for the same period
The 'Closing Entry' button in the Tax Report has inconsistent behavior when the closing entry is already posted and then reset to draft. In such cases, the button points to an empty list of closing entries instead of opening the existing one. Additionally, there is a possibility of multiple closing entries being created for the same period, leading to incorrect reporting.

To address these issues, the following changes have been implemented:

1. When clicking on the 'Closing Entry' button:
- If the closing entry already exists for the period, it is opened.
- If the closing entry doesn't exist for the period, a new closing entry is created and opened, regardless of the tax lock date.

2. A check has been added to prevent multiple closing entries for the same period. If a closing entry already exists in draft state for the given period, it is updated with the current computations instead of creating a new entry.

3. If a closing entry is already posted but then reset to draft (where the tax lock date is greater than or equal to the end date of the closing entry period), the button will still open the existing entry.

closes odoo/enterprise#41562

Task-id: 3284543
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>