[ADD] {website_}appointment{_account_payment,_sale}: add payment to appointments
This commit adds modules to allow configuration of an up-front payment for
appointments as well as the payment flow on the front-end. When eCommerce
is not installed, it uses account_payment tools to manage payment. Otherwise,
it is fully managed by classical eCommerce Flow.

*** APPOINTMENT_ACCOUNT_PAYMENT ***

You can now configure a product on appointment_types, of type 'booking_fees'.
This will be the product to buy when booking the appointment. This is a
single 'booking fee' (not per capita atm).

The calendar event linked to the booking should NOT be created until it is paid.
We want to avoid synchronizing calendar events that are not really settled in
terms of payment. Therefore, two new models are created to carry all necessary
booking data. We try to avoid new fields on the calendar.event and to keep new
stored fields introduced in the feature on those new models.

- CALENDAR.BOOKING: counterpart to calendar event, carrying all information linked
to an appointment booking process.
- CALENDAR.BOOKING.LINE: counterpart to appointment.booking.line, this is used
only to store resource booking information (for instance, booking 3 tables adding
up to a certain capacity...)

(Those are technical models and should not be manually updated in most cases)
Availabilities will NOT be modified according to these models, as their existence
should be equivalent to a 'pending' booking, as long as no calendar_event_id is
linked to them. A not paid booking does not reserve any space.

FLOW:
    1) On Appointment Form Submission: we create a calendar.booking and link it
to the booking. An invoice should only be linked to one booking, in this flow.
    2) On invoice payment => create calendar.event from booking values

When do we trigger a 'paid' invoice? On _invoice_paid_hook: this hook is called
on both backend account.payment registration and successful transaction as well
from the payment flow.

    And what about collisions?

If we cannot create an event because the user / resource are not available or
configured correctly anymore, we do not create the event, and log a link to
the form view of failed bookings on the invoice. Manual action will be needed.

A check method _filter_unavailable_bookings is callable on a recordset of
bookings and will try to fit the most bookings according to user / resource
availability. It returns the other ones. This method should be used whenever
an availability check is necessary on bookings. (*)

    How does the front-end payment flow work?

We override payment_pay rendering values to add custom routes and design. We add
a simplified custom payment screen as well. We specify custom cancel and landing
routes. The rest is handled by account_payment and payment tools to pay an
invoice with an access_token, using invoice routes.

On cancelling, the user will have to take another appointment and create another
booking (on a new invoice) The invoice posting is handled by providers and
transaction flow (see existing account_payment controllers)

The landing route is still the page of calendar.event when successfully paid, and
a similar page for the booking otherwise. A new template is added to do so.

    When are bookings removed?

They are collected and removed by the garbage collector, if its stop datetime
was at least 2 months ago, or if created 6 at least months ago. Payments should
be done in that time window.

*** WEBSITE_APPOINTMENT_ACCOUNT_PAYMENT ***

Just a small bridge module to publish data, and payment chevron and update ui
to fit what exists already in website_appointment

*** WEBSITE_APPOINTMENT_SALE ***

When eCommerce is installed, its flow takes over completely. Now, when creating
the booking, we link it to a sale order line. This means someone can book
different slots or appointments in the same cart. We use the request session
if available as we want the description tz and dts to match the one picked
by the user, if any. The lang is the order partner's as per usual sale flow.

When do we trigger a 'paid' SO? We consider that we can do so as soon as
the Sale order is confirmed. This is a bottleneck that is sufficient for
both the case of successful front-end transaction and backend actions.

FLOW:
    1) On appointment form submission, adds a line to cart linked to new booking.
    2) On SO confirm: create calendar.event from booking values

    And what about collisions?

When trying to pay the cart, a check is done to make sure booked slots are
still available (*). Quantity is always set to 1 as fee is unique per booking.
We also log links to failed bookings on SO.

    Confirmation Page

We add a card on SO confirmation page for events and bookings, with appropriate
status, date and apt intro msg, as well as a link to the calendar.event/booking
page depending on whether the SO was confirmed through the transaction and
calendar event was successfully created. For instance if SO had two lines (-> two
bookings), one failed due to collision, one succeeded, then we have one card of
a calendar event, and one of calendar booking (mostly the same)

OTHER CHANGES

- phone is now used in place of mobile in appointment booking flow.
- controllers of appointment (submission...) are restructured for easy overrides.
- constraint added to have either calendar event or booking on answer inputs.

Task-3079302
UPG PR: odoo/upgrade#5289

closes odoo/enterprise#46891

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>