[FIX] account_reports: prevent traceback when tax_closing_end_date field is null
While creating a Journal Entries , if the value of 'tax_closing_end_date' field
is passed as null. As 'tax_closing_end_date' field is not required, there might
be a case where it can be null/False.

Traceback will be generated.

unsupported operand type(s) for +: 'bool' and 'relativedelta'

This commit will resolve this issue.

sentry-4212662929

closes odoo/enterprise#41622

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>