[IMP] account_sepa: handle sepa structured reference
When creating SEPA files to perform payments, we did not check for
structured references and always put the reference in the `Ustrd` field
in the SEPA xml file.

This change makes sure we check whether the reference of a move is
a structured reference (ISO CF, BE, CH, FI, NO or SE), and if so, add
the reference in the right `Strd` field along with the correct tags
for that type of structured reference.

task-3284330

closes odoo/enterprise#41629

Related: odoo/odoo#122835
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>