[FIX] delivery_fedex: send dimensions as integer
To reproduce: set up FedEx shipping method, use 'FEDEX_YOUR_PACKAGING'
as the package type. Make a sale order and attempt to get the shipping
rate for FedEx. It will lead to the following error in the logs:

```
validation failure for RateRequest Error:cvc-datatype-valid.1.2.1:
'5.0' is not a valid value for 'integer'.
cvc-type.3.1.3: The value '5.0' of element 'ns0:Length' is not valid.
cvc-datatype-valid.1.2.1: '5.0' is not a valid value for 'integer'.
cvc-type.3.1.3: The value '5.0' of element 'ns0:Width' is not valid.
cvc-datatype-valid.1.2.1: '5.0' is not a valid value for 'integer'.
cvc-type.3.1.3: The value '5.0' of element 'ns0:Height' is not valid.
```

From saas-16.1, the dimensions of the packages are stored as float. To
fix this issue we need to send them as integer.

opw-3330372

closes odoo/enterprise#41666

X-original-commit: 6f349803f2d0d4f60dc3d52bc767809c13389555
Signed-off-by: Tiffany Chang <tic@odoo.com>