[IMP] project: rename documents workspace when renaming a project
Before this commit when we create a project then workspace was set as
project/project name but when we rename a project the workspace was not renamed.

After this commit if the workspace is not linked with other projects then it will
be modified if we modify/rename the project name.

task-3330282

closes odoo/enterprise#41691

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>