[FIX] barcode: parse number >= 1000 formatted is wrong
In https://github.com/odoo/enterprise/commit/4da8caacc4ecd69b20b96464721e4fac6e1193b3,
the fix done did not take the formatting of numbers bigger (or equal)
than 1000 into account. For example, 1000 become "1.000,00".

To fix this, simply modified the `thousandsSep` to an empty string and the decimalPoint to a point.

closes odoo/enterprise#42911

X-original-commit: 955899c6922d15094e4458777ff1cf3966e89fa1
Signed-off-by: Steve Van Essche <svs@odoo.com>