[FIX] timesheet_grid: Fix timer `z-index`
=== ISSUE ===

If you navigate to Timesheet > Click on Start > Select an item from the
dropdown, the dropdown menu is hidden behind
the`.o_grid_navigation_wrap` item.

This issue is due to the fact that `.o_grid_navigation_wrap` uses
the `o-gantt-zindex`to set its z-index, while the `.timesheet-timer`
menu does not have a `z-index` set.

=== AFTER ===

We set a `z-index` higher than the `.o_grid_navigation_wrap`, using
the same mixin to keep consistency. We also set a `position-relative` in
the XML to be able to set the `z-index` correctly.

task-3344740
part of task-3326263

closes odoo/enterprise#42645

X-original-commit: 5a318b1af6a19a0931c9ee6c038d8e3a1c49ba89
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>