[FIX] web: superuser background-color
=== ISSUE ===

Before we introduced the Milk new design, the navbar component used
to have a background-color based on the $o-brand-odoo SCSS variable.
This variable would use the primary color on both Community
and Enterprise.

With the introduction of Milk, the UI received a fresh upgrade including
(in our case) a white navbar rather than using the primary-color.

=== AFTER ===

Since we cannot rely on the $o-brand-odoo to make the component anatomy
consistent across Community/Enterprise and the darkmode, we refactored
the navbar using the `border-image` CSS property and adjust the height
of the menu to fit its content.

task-3337789
part of task-332626

closes odoo/enterprise#41916

X-original-commit: 86f937130d2a6cebd774c69b1094696a9e0a9c88
Related: odoo/odoo#123613
Signed-off-by: Pierre Paridans (app) <app@odoo.com>